chore: fix all lint warnings and ignore build artifacts
Ignore src-tauri/target and backend-sidecar in eslint so Tauri/Rust
build output stops producing tens of thousands of parse errors. Then
fix the real warnings at the root cause: memoize workspaces arrays,
capture refs for cleanup, add missing queryClient deps, extract
non-component exports into sibling modules to satisfy react-refresh.
