+++ frontend/src/utils/platform.ts
@@ -1,0 +1,1 @@
+export const IS_MAC_PLATFORM = navigator.platform.toUpperCase().startsWith('MAC');
